[Commit] Iteration 437 #errors = 22 <= 22.
Addressing `BuildError `[/tmp/tempfspartition/QnetTransformArtifacts/MidTransformCode/sourceCode/MvcAppToModernize/WebApplication/Controllers/ProductController.cs][CS0518][44,55] Predefined type 'System.Int32' is not defined or imported``.

+++ MvcAppToModernize/WebApplication/Controllers/ProductController.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 
 namespace WebApplication.Controllers
 {
+using System;
     public class ProductViewModel
     {
         public IEnumerable<Product> Products { get; set; }
@@ -41,7 +42,7 @@ namespace WebApplication.Controllers
             var viewModel = new ProductViewModel
             {
                 Products = products,
-                CartItemCount = cartItems?.Sum(c => c.Quantity) ?? 0
+                CartItemCount = cartItems?.Sum(c => (System.Int32)c.Quantity) ?? 0
             };
 
             return Ok(viewModel);
